PLAT-781: put code to possibly right place
